tailscale-operator/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-kubernetes>=4.0.0,<5.0.0
python-dotenv==1.0.0